Deleted Location package, added attributes to the Airport class
